+++ src/onegov.user/onegov/user/__init__.py
@@ -1,3 +1,10 @@
+import logging
+log = logging.getLogger('onegov.user')  # noqa
+log.addHandler(logging.NullHandler())  # noqa
+
+from translationstring import TranslationStringFactory
+_ = TranslationStringFactory('onegov.user')  # noqa
+
 from onegov.user.collection import UserCollection
 from onegov.user.model import User