+++ src/onegov.user/onegov/user/forms/__init__.py
@@ -1,0 +1,3 @@
+from onegov.user.forms.login import LoginForm
+
+__all__ = ['LoginForm']