+++ src/onegov.user/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Adds a generic login form
+  [href]
+
 0.0.2 (2015-06-26)
 ~~~~~~~~~~~~~~~~~~~